+++ src/main/java/com/easypost/model/BetaPaymentRefund.java
@@ -1,0 +1,12 @@
+package com.easypost.model;
+
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class BetaPaymentRefund extends EasyPostResource{
+    private String refundedAmount;
+    private String paymentLogId;
+    private List<String> refundedPaymentLogs;
+    private List<Error> errors;
+}